# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API - Swagger Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login A User Account",
                "operationId": "b9234032ffcfcb60d967448ab7463927",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "View A User Account",
                "operationId": "e6090d1b236d51807a6f7c53b5b5da23",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/create": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create A New User Account",
                "operationId": "81dab8fa48f81cd38b98e656cf1c99fc",
                "requestBody": {
                    "description": "Supply User Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUserProperties"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/user/orders": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List All User Orders",
                "operationId": "ea7817bd2daa7204c2175dfc02960767",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page Number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Item Per page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort By",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "Direction of sort",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/edit": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update A User Account",
                "operationId": "04fea39db5e6b52731205f58a3c437b7",
                "requestBody": {
                    "description": "Update User Record",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUserProperties"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/user/forgot-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Request Forgot Password Token",
                "operationId": "75b0b5e934c64b70dcca82226f9ff0fc",
                "requestBody": {
                    "description": "Pass user email",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@email.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/user/reset-password-token": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Recover User Account",
                "operationId": "fcf8706b007e09d6fb2d0be8d3ea1a4d",
                "requestBody": {
                    "description": "Supply Credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@email.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "format": "token",
                                        "example": "gfgfg885478hgf"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Logout A User Account",
                "operationId": "a6d081124274ad86c6409c5bf1c0dfb8",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "RegisterUserProperties": {
                "required": [
                    "email",
                    "password",
                    "last_name",
                    "first_name",
                    "address",
                    "phone_number"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "format": "first_name",
                        "example": "first_name"
                    },
                    "last_name": {
                        "type": "string",
                        "format": "last_name",
                        "example": "last_name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@email.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    },
                    "avatar": {
                        "type": "string",
                        "format": "avatar",
                        "example": "avatar"
                    },
                    "address": {
                        "type": "string",
                        "format": "address",
                        "example": "address"
                    },
                    "phone_number": {
                        "type": "string",
                        "format": "phone_number",
                        "example": "phone_number"
                    },
                    "is_marketing": {
                        "type": "string",
                        "format": "is_marketing",
                        "example": "is_marketing"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RegisterUserProperties"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "scheme": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}